MAJ de la manière de gérer les doublons lors de l'import et non suppression d'un élément utilisé (source, catégorie, type)
git-svn-id: svn://svn.noparking.net/usr/local/svn/lozeil@49 d07dd117-7d15-4c26-aaaa-e67aa2e0e813

+++ contents/categories.php
@@ -25,12 +25,11 @@ if (isset($_POST['submit'])) {
 		if ($category->name != $name and !empty($name)) {
 			$category->name = $name;
 			$category->save();
-		} elseif (empty($name)) {
+		} elseif (empty($name) and $category->is_deletable()) {
 			$category->delete();
 		}
 	}
 }
-
 
 $menu = new Menu_Area();
 $menu->prepare_navigation(__FILE__);
